Extracts envelop logic to its own file
