+++ src/api/graphql.ts
@@ -1,24 +1,7 @@
-/* eslint-disable react-hooks/rules-of-hooks */
 import type { GatsbyFunctionRequest, GatsbyFunctionResponse } from 'gatsby'
-import type { FormatErrorHandler } from '@envelop/core'
-import {
-  enableIf,
-  envelop,
-  useSchema,
-  useMaskedErrors,
-  useExtendContext,
-} from '@envelop/core'
-import { GraphQLError } from 'graphql'
-import { useGraphQlJit } from '@envelop/graphql-jit'
-import { useParserCache } from '@envelop/parser-cache'
-import { useValidationCache } from '@envelop/validation-cache'
-import { useResponseCache } from '@envelop/response-cache'
 
-import { getSchema, getContextFactory } from '../server'
+import { enveloped, isBadRequestError } from '../server/envelop'
 import persisted from '../../@generated/graphql/persisted.json'
-
-const { NODE_ENV, CONTEXT: ENV = NODE_ENV } = process.env
-const isProduction = ENV === 'production'
 
 const persistedQueries = new Map(Object.entries(persisted))
 
@@ -44,37 +27,6 @@ const parseRequest = (req: GatsbyFunctionRequest) => {
   }
 }
 
-const isBadRequestError = (err: GraphQLError) => {
-  return err.originalError && err.originalError.name === 'BadRequestError'
-}
-
-const maskError: FormatErrorHandler = (err: GraphQLError | unknown) => {
-  if (err instanceof GraphQLError) {
-    if (!isBadRequestError(err)) {
-      return new GraphQLError('Sorry, something went wrong.')
-    }
-
-    return err
-  }
-
-  return new GraphQLError('Sorry, something went wrong.')
-}
-
-const createGetEnveloped = async () =>
-  envelop({
-    plugins: [
-      useSchema(await getSchema()),
-      useExtendContext(getContextFactory()),
-      enableIf(isProduction, () => useMaskedErrors({ formatError: maskError })),
-      useGraphQlJit(),
-      useValidationCache(),
-      useParserCache(),
-      useResponseCache(),
-    ],
-  })
-
-const enveloped = createGetEnveloped()
-
 const handler = async (
   req: GatsbyFunctionRequest,
   res: GatsbyFunctionResponse
@@ -85,7 +37,7 @@ const handler = async (
     return
   }
 
-  const getEnveloped = await enveloped
+  const getEnveloped = await enveloped()
   const { parse, contextFactory, execute, schema } = getEnveloped({ req })
   const { operationName, variables, query } = parseRequest(req)
